service-workers/service-worker/sandboxed-iframe-by-csp-fetch-event.https.html

<!DOCTYPE html>
<title>ServiceWorker FetchEvent for iframe that is sandboxed via CSP HTTP response header.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
  <script>
  let lastCallbackId = 0;
  let callbacks = {};
  function postMessageAndWaitResult(frame, type) {
    return new Promise(resolve => {
      let id = ++lastCallbackId;
      callbacks[id] = resolve;
      frame.contentWindow.postMessage({id: id, type: type}, '*');
    });
  }

  function postMessageAndGetRequestURLs(worker) {
    return new Promise(resolve => {
      let channel = new MessageChannel();
      channel.port1.onmessage = msg => {
        resolve(msg.data.requests);
      };
      worker.postMessage({port: channel.port2}, [channel.port2]);
    });
  }

  window.onmessage = function (e) {
    message = e.data;
    let id = message['id'];
    let callback = callbacks[id];
    delete callbacks[id];
    callback(message['result']);
  };

  const SCOPE = 'resources/sandboxed-iframe-fetch-event-iframe.py';
  const SCRIPT = 'resources/sandboxed-iframe-fetch-event-worker.js';
  const expected_base_url = new URL(SCOPE, location.href);
  // A service worker controlling |SCOPE|.
  let worker;
  // An iframe whose response header has
  // 'Content-Security-Policy: allow-scripts'.
  // This should NOT be controlled by a service worker.
  let sandboxed_non_same_origin_frame;
  // An iframe whose response header has
  // 'Content-Security-Policy: allow-scripts allow-same-origin'.
  // This should be controlled by a service worker.
  let sandboxed_same_origin_frame;

  promise_test(t => {
    const scope_full_path = new URL(SCOPE, location.href).href;
    return service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
        add_completion_callback(() => registration.unregister());
        worker = registration.installing;
        return wait_for_state(t, registration.installing, 'activated');
      })
      .then(() => with_iframe(SCOPE + '?sandbox=allow-scripts&script-header'))
      .then(f => {
        sandboxed_non_same_origin_frame = f;
        add_completion_callback(() => f.remove());
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 1);
        assert_equals(requests[0],
                      scope_full_path + '?sandbox=allow-scripts&script-header');
      })
      .then(() => with_iframe(
        SCOPE + '?sandbox=allow-scripts allow-same-origin&script-header'))
      .then(f => {
        sandboxed_same_origin_frame = f;
        add_completion_callback(() => f.remove());
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 1);
        assert_equals(requests[0],
                      scope_full_path + '?sandbox=allow-scripts%20allow-same-origin&script-header');
      })
  }, 'Preparation of a service worker and iframes in the scope of the ' +
     'service worker.');

  promise_test(t => {
    let frame = sandboxed_non_same_origin_frame;
    return postMessageAndWaitResult(frame, 'fetch')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 0,
                      'The request should NOT be handled by SW.');
      });
  }, 'Fetch request from iframe sandboxed by CSP HTTP header with ' +
     'allow-scripts flag');

  promise_test(t => {
    let frame = sandboxed_non_same_origin_frame;
    return postMessageAndWaitResult(frame, 'workerfetch')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 0,
                      'The request should NOT be handled by SW.');
      });
  }, 'Fetch request from a worker in iframe sandboxed by CSP HTTP header ' +
     'allow-scripts flag');

  promise_test(t => {
    let frame = sandboxed_non_same_origin_frame;
    return postMessageAndWaitResult(frame, 'iframe')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 0,
                      'The request should NOT be handled by SW.');
      });
  }, 'Request for an iframe in the iframe sandboxed by CSP HTTP header with ' +
     'allow-scripts flag');

  promise_test(t => {
    let frame = sandboxed_non_same_origin_frame;
    return postMessageAndWaitResult(frame, 'script')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 0,
                      'The request should NOT be handled by SW.');
      });
  }, 'Request for an sandboxed iframe with allow-scripts flag in the iframe ' +
     'sandboxed by CSP HTTP header with allow-scripts flag');

  promise_test(t => {
    let frame = sandboxed_non_same_origin_frame;
    return postMessageAndWaitResult(frame, 'script-origin')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 0,
                      'The request should NOT be handled by SW.');
      });
  }, 'Request for an sandboxed iframe with allow-scripts and ' +
     'allow-same-origin flag in the iframe sandboxed by CSP HTTP header with ' +
     'allow-scripts flag');

  promise_test(t => {
    let frame = sandboxed_same_origin_frame;
    return postMessageAndWaitResult(frame, 'fetch')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 1,
                      'The request should be handled by SW.');
        assert_equals(requests[0], expected_base_url + '?test=fetch');
      });
  }, 'Fetch request from iframe sandboxed by CSP HTTP header with ' +
     'allow-scripts and allow-same-origin flag');

  promise_test(t => {
    let frame = sandboxed_same_origin_frame;
    return postMessageAndWaitResult(frame, 'workerfetch')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 1,
                      'The request should be handled by SW.');
        assert_equals(requests[0], expected_base_url + '?test=workerfetch');
      });
  }, 'Fetch request from a worker in iframe sandboxed by CSP HTTP header ' +
     'with allow-scripts and allow-same-origin flag');

  promise_test(t => {
    let frame = sandboxed_same_origin_frame;
    return postMessageAndWaitResult(frame, 'iframe')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 1,
                      'The request should be handled by SW.');
        assert_equals(requests[0], expected_base_url + '?test=iframe');
      });
  }, 'Request for an iframe in the iframe sandboxed by CSP HTTP header with ' +
     'allow-scripts and allow-same-origin flag');

  promise_test(t => {
    let frame = sandboxed_same_origin_frame;
    return postMessageAndWaitResult(frame, 'script')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 0,
                      'The request should NOT be handled by SW.');
      });
  }, 'Request for an sandboxed iframe with allow-scripts flag in the ' +
     'iframe sandboxed by CSP HTTP header with allow-scripts and ' +
     'allow-same-origin flag');

  promise_test(t => {
    let frame = sandboxed_same_origin_frame;
    return postMessageAndWaitResult(frame, 'script-origin')
      .then(result => {
        assert_equals(result, 'done');
        return postMessageAndGetRequestURLs(worker);
      })
      .then(requests => {
        assert_equals(requests.length, 1,
                      'The request should be handled by SW.');
        assert_equals(requests[0], expected_base_url + '?test=script-origin');
      });
  }, 'Request for an sandboxed iframe with allow-scripts and ' +
     'allow-same-origin flag in the iframe sandboxed by CSP HTTP header with ' +
     'allow-scripts and allow-same-origin flag');
  </script>
</body>
